requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
apscheduler>=3.10.0
pandas>=2.0.0
numpy>=1.23.2
matplotlib>=3.7.0
alpaca-trade-api>=3.0.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
mdb-engine>=0.2.3
pymongo>=4.6.0
python-multipart>=0.0.6
pydantic>=2.0.0
jinja2>=3.1.0
firecrawl-py>=0.0.16
vectorbt>=0.25.0
pyarrow>=14.0.0
